[Ref] Sample Converters: Fix clang -Wdouble-promotion warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@9179 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundbase/SampleFormatConverters.h
@@ -797,7 +797,7 @@ struct Convert<double, float>
 	typedef double output_t;
 	MPT_FORCEINLINE output_t operator() (input_t val)
 	{
-		return val;
+		return static_cast<double>(val);
 	}
 };
